% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIR.Euler.sim.R
\name{CIR.Euler.sim}
\alias{CIR.Euler.sim}
\title{CIR Euler-Maruyama discretization approach.}
\usage{
CIR.Euler.sim(
  t0 = 0,
  T = 1,
  Delta = 1/252,
  r0 = 0.03,
  gamma = 0.3807,
  rbar = 0.072,
  alpha = 0.0548,
  M = 10
)
}
\arguments{
\item{t0}{starting point of interest rate path.}

\item{T}{ending time point.

both t0 and T are given in years.}

\item{Delta}{is the time step; default value being 1 day (1/252 in years).}

\item{r0}{is the starting value of the path}

\item{gamma}{is the mean reversion speed}

\item{rbar}{is the long term mean}

\item{alpha}{rbar*gamma> alpha/2 not to have negative rates}

\item{M}{is the number of paths to be simulated
Default values of r0, gamma, rbar and alpha are the ones given in Veronesi(2010) page 552 footnote 10.}
}
\description{
This function simulate M  interest rate paths from the Cox-Ingersoll-Ross model using Euler-Maruyama discretization approach
}
