% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIR.gmm.R
\name{CIR.gmm}
\alias{CIR.gmm}
\title{Function to calculate objective function in GMM minimization for CIR model.}
\usage{
CIR.gmm(param, X)
}
\arguments{
\item{param}{is c(alpha1,beta1,sigma)
\deqn{\alpha_1  = \gamma\bar{r}\Delta}
\deqn{\beta_1 = 1- \gamma\Delta}
\deqn{\sigma = \sqrt{\alpha\Delta}}}

\item{X}{contains a sample path}
}
\description{
\deqn{f_1 = \frac{1}{N}\sum_{i=1}^N \left(r(i)- \alpha_1-\beta_1r(i-1)\right)}
\deqn{f_2 = \frac{1}{N}\sum_{i=1}^N [\left(r(i)- \alpha_1-\beta_1r(i-1)\right)^2 -r(i-1)\sigma^2]}
\deqn{ f_3 = \frac{1}{N}\sum_{i=1}^N \left[r(i)- \alpha_1-\beta_1r(i-1)\right]r(i-1)}
}
