% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIR.J.R
\name{CIR.J}
\alias{CIR.J}
\title{sum of squared difference between market prices and theoretical prices of zeros in CIR.}
\usage{
CIR.J(param, r0 = 0.0168, alpha = NULL, bond.prices, bond.maturities)
}
\arguments{
\item{param}{is a list of parameters c(\eqn{\gamma,\bar{r},\alpha}).}

\item{r0}{initial short rate default value is 0.0168}

\item{alpha}{if it is set to null minimization is  with respect to all three 
parameters \eqn{\gamma,\bar{r}} and \eqn{\alpha} otherwise
its minimiztion is with respect to \eqn{\gamma} and \eqn{\bar{r}} only.}

\item{bond.prices}{a list of observed zero coupon bond prices (or discount factors) 
for bond.maturities}

\item{bond.maturities}{a list bond maturities}
}
\description{
Function to compute the sum of squared of differences between the CIR theoretical bond price
and the market price of zero-coupon bonds.
This function will be used in ``nlm" and ``optim" as the objective function to be minimized.
It uses the function CIR.zcbp
}
