% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIR.log.density.R
\name{CIR.log.density}
\alias{CIR.log.density}
\title{Function to calculate log of probability density function.}
\usage{
CIR.log.density(x, Delta, x0, gamma, rbar, alpha)
}
\arguments{
\item{x}{next observation}

\item{Delta}{is the time step; default value being 1 day (1/252 in years).}

\item{x0}{current observation}

\item{gamma}{is the mean reversion speed}

\item{rbar}{is the long term mean}

\item{alpha}{rbar*gamma> alpha/2 not to have negative rates}
}
\description{
It calculates the log of the following density function.
\deqn{f(r_{t+s}|r_t) = c_s \chi^2(c_sr_{t+s},\nu,\lambda_{t+s})}
where \eqn{\chi^2(.,\nu,\lambda_{t+s})} is a noncentral \eqn{\chi^2} density function  with \eqn{\nu} degree of freedom,
and non centrality parameter \eqn{\lambda_{t+s}}, with
\deqn{c_s = \frac{4\gamma}{\alpha(1-\exp(-\gamma s))}}
\deqn{\nu = \frac{4\gamma}{\alpha} \bar{r}}
\deqn{\lambda_{t+s} = c_sr_t \exp(-\gamma s)}
}
