% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIR.log.lik.R
\name{CIR.log.lik}
\alias{CIR.log.lik}
\title{Log likelihood function of CIR model.}
\usage{
CIR.log.lik(param, X)
}
\arguments{
\item{param}{is the list  c(gamma, rbar,alpha)}

\item{X}{contains an interest rate path a time series object}
}
\description{
It takes one interest rate path and calculate log likelihood function.
}
\details{
It uses the function CIR.log.density.
}
