% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIR.neg.prob.R
\name{CIR.neg.prob}
\alias{CIR.neg.prob}
\title{Function to calculate CIR negative probabilities}
\usage{
CIR.neg.prob(
  s = 1/252,
  rt = 0.2/100,
  gamma = 0.3262,
  rbar = 0.07,
  alpha = 0.0584
)
}
\arguments{
\item{s}{is the time step; default value being 1 day (1/252 in years).}

\item{rt}{is the current value}

\item{gamma}{is the mean reversion speed}

\item{rbar}{is the long term mean}

\item{alpha}{rbar*gamma> alpha/2 not to have negative rates}
}
\description{
This function calculates the probability of next value become negative in the 
CIR path under Euler discretization.
It calculates conditional probability \eqn{P[r_{t+s}<0|r_t]}
}
