% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIR.yield.R
\name{CIR.yield}
\alias{CIR.yield}
\title{Yield of a zero coupon bond in  CIR model}
\usage{
CIR.yield(
  r0 = 0.02,
  t = 0,
  T = 1,
  gamma = 0.3807,
  rbar = 0.072,
  alpha = 0.0548
)
}
\arguments{
\item{r0}{is the short rate at time t}

\item{t}{time where zero coupon bond price is evaluated in years.}

\item{T}{maturity time on years.}

\item{gamma}{is the mean reversion speed}

\item{rbar}{is the long term mean}

\item{alpha}{rbar*gamma> alpha/2 not to have negative rates}
}
\description{
It implements the formulas in the Section 3.5 of the article.
\deqn{ B(t;T) = \frac{2\left( e^{\psi_1 (T-t)} - 1\right)}{(\gamma^* + \psi_1)\left(e^{\psi_1(T-t)}-1\right) + 2\psi_1}}
\deqn{ A(t;T) = 2\frac{\bar{r}^* \gamma^*}{\alpha}\log \left( \frac{2\psi_1  e^{(\psi_1+\gamma^*)\frac{(T-t)}{2}} }{(\gamma^* + \psi_1)\left(e^{\psi_1(T-t)}-1\right) + 2\psi_1} \right)}
\deqn{\psi_1 = \sqrt{(\gamma^*)^2+2\alpha}  }
\deqn{Z(r,t;T) = e^{A(t;T)-B(t;T)\times r}}
\deqn{Yield = -\frac{ ln(Z(r,t:T))}{T-t} }
}
