% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIR.zcbp.R
\name{CIR.zcbp}
\alias{CIR.zcbp}
\title{Zero coupon bond prices under CIR model}
\usage{
CIR.zcbp(r0 = 0.02, t = 0, T = 1, gamma = 0.3807, rbar = 0.072, alpha = 0.0548)
}
\arguments{
\item{r0}{is the short rate at time t}

\item{t}{time where zero coupon bond price is evaluated}

\item{T}{maturity time

both t and T are given in years.}

\item{gamma}{is the mean reversion speed}

\item{rbar}{is the long term mean}

\item{alpha}{rbar*gamma> alpha/2 not to have negative rates}
}
\description{
It implements the CIR zero-coupon bond pricing formual given in the Section 3.5 of the article.
\deqn{ B(t;T) = \frac{2\left( e^{\psi_1 (T-t)} - 1\right)}{(\gamma^* + \psi_1)\left(e^{\psi_1(T-t)}-1\right) + 2\psi_1}}
\deqn{\qquad A(t;T) = 2\frac{\bar{r}^* \gamma^*}{\alpha}\log \left( \frac{2\psi_1  e^{(\psi_1+\gamma^*)\frac{(T-t)}{2}} }{(\gamma^* + \psi_1)\left( e^{\psi_1(T-t)}-1\right) + 2\psi_1} \right)}
\deqn{\psi_1 = \sqrt{(\gamma^*)^2+2\alpha}  }
\deqn{Z(r,t;T) = e^{A(t;T)-B(t;T)\times r}   }
}
