% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ho.Lee.call.zcb.R
\name{Ho.Lee.call.zcb}
\alias{Ho.Lee.call.zcb}
\title{price of  a call option on a zero-coupon bond under Ho-Lee model.}
\usage{
Ho.Lee.call.zcb(
  TO = 1,
  TB = 5,
  K = 0.8,
  sigma = 0.0221,
  ZCBTB = 0.7978,
  ZCBTO = 0.979
)
}
\arguments{
\item{TO}{option maturity}

\item{TB}{maturity of underlying bond}

\item{K}{strike price of the option}

\item{sigma}{instantaneous Ho-Lee volatility}

\item{ZCBTB}{bond price of a zero coupon that matures at time TB}

\item{ZCBTO}{bond price of a zero coupon that matures at time TO}
}
\description{
price of  a call option on a zero-coupon bond under Ho-Lee model.
}
