% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ho.Lee.cap.price.vector.R
\name{Ho.Lee.cap.price.vector}
\alias{Ho.Lee.cap.price.vector}
\title{Calculate cap price vector.}
\usage{
Ho.Lee.cap.price.vector(
  cap_rates,
  Maturity,
  Delta = 0.25,
  discount_factors,
  sigma
)
}
\arguments{
\item{cap_rates}{cap rates starting at 0.25 going upto the largest cap}

\item{Maturity}{vector containing 0.25,0.5,0.75 etc}

\item{Delta}{reciprocal of number of payments per year; it is assumed 4 times payments, Delta =0.25}

\item{discount_factors}{for times 0.25, 0.5,0.75...}

\item{sigma}{Ho-lee  instantaneous volatility which is a constant}
}
\description{
This calculates a vector of cap prices and it uses Ho.Lee.cap function.
}
