% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ho.Lee.cap.R
\name{Ho.Lee.cap}
\alias{Ho.Lee.cap}
\title{Price of a cap.}
\usage{
Ho.Lee.cap(rK = 0.05, TK = 5, Delta = 0.25, discount_factors, sigma)
}
\arguments{
\item{rK}{cap rate
first caplet payment starts at time 2*Delta
discount_curve must have discount factors for time points: Delta, 2*Delta,3*Delta,...T2
calculate values of all the caplet that makes payment from 2*Delta to T2 and add them up}

\item{TK}{is the last caplet payment.}

\item{Delta}{reciprocal of number of payments pere year; it is assumed 4 times payments, Delta =0.25}

\item{discount_factors}{for times 0.25, 0.5,0.75...}

\item{sigma}{Ho-lee  instantaneous volatility which is a constant}
}
\description{
It is  calculated as a sum of individual caplet prices.
}
