% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hull.White.call.zcb.R
\name{Hull.White.call.zcb}
\alias{Hull.White.call.zcb}
\title{Price of a call option on a zero-coupon bond under Hull-White model.}
\usage{
Hull.White.call.zcb(
  TO = 1,
  TB = 5,
  K = 0.8,
  gamma = 0.019,
  sigma = 0.0196,
  ZCBTB = 0.7978,
  ZCBTO = 0.979
)
}
\arguments{
\item{TO}{option maturity}

\item{TB}{maturity of underlying bond}

\item{K}{strike price of the option}

\item{gamma}{mean reversion rate of Hull-White}

\item{sigma}{instantaneous Hull-White volatility}

\item{ZCBTB}{bond price of a zero coupon that matures at time TB}

\item{ZCBTO}{bond price of a zero coupon that matures at time TO}
}
\description{
It is almost similar to call option price under Vasicek model, the differenec is market prices of zero-copon binds are used instead of
model prices.
}
