% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hull.White.cap.price.vector.R
\name{Hull.White.cap.price.vector}
\alias{Hull.White.cap.price.vector}
\title{Calculate cap price vector for Hull-White model.}
\usage{
Hull.White.cap.price.vector(
  cap_rates,
  Maturity,
  Delta = 0.25,
  discount_factors,
  gamma,
  sigma
)
}
\arguments{
\item{cap_rates}{cap rates starting at 0.25 going up to the largest cap}

\item{Maturity}{vector containing 0.25,0.5,0.75 etc}

\item{Delta}{reciprocal of number of payments per year; it is assumed 4 times payments, Delta =0.25}

\item{discount_factors}{for times 0.25, 0.5,0.75...}

\item{gamma}{speed of mean reversion}

\item{sigma}{Hull-White instantaneous volatility which is constant}
}
\description{
This function uses the function Hull.White.cap.
}
\details{
Parameters are identical to that of cap price in the same model.
}
