% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hull.White.cap.price.vectorv2.R
\name{Hull.White.cap.price.vectorv2}
\alias{Hull.White.cap.price.vectorv2}
\title{Calculate cap price vector}
\usage{
Hull.White.cap.price.vectorv2(
  cap_rates,
  Maturity,
  Delta = 0.25,
  discount_factors,
  sigma
)
}
\arguments{
\item{cap_rates}{cap rates starting at 0.25 going up to the largest cap}

\item{Maturity}{vector containing 0.25,0.5,0.75 etc}

\item{Delta}{reciprocal of number of payments per year; it is assumed 4 times payments, Delta =0.25}

\item{discount_factors}{for times 0.25, 0.5,0.75...}

\item{sigma}{Hull-White forward volatility applicable to each point in Maturity}
}
\description{
The difference between this function and Hull.White.cap.price.vector is that this function uses something similar to 
forward volatilities.
}
