% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hull.White.caplet.R
\name{Hull.White.caplet}
\alias{Hull.White.caplet}
\title{Price of a caplet  with Hull-White model.}
\usage{
Hull.White.caplet(
  rK = 0.03815,
  T = 5,
  Delta = 0.25,
  ZT1 = 0.835326,
  ZT2 = 0.8247441,
  gamma = 0.054523,
  sigma = 0.0149
)
}
\arguments{
\item{rK}{cap rate.}

\item{T}{is the time of caplet payment.}

\item{Delta}{is the spacing of payments, normally its 0.25}

\item{ZT1}{is the discount factor for time T-Delta.}

\item{ZT2}{is the discount factor for time T.}

\item{gamma}{Hull-White gamma.}

\item{sigma}{Hull-White sigma.}
}
\description{
It is based on that fact that caplet price can be viewed as put option on a zero-coupon bond.
}
