% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hull.White.capletv2.R
\name{Hull.White.capletv2}
\alias{Hull.White.capletv2}
\title{Price of a caplet v2}
\usage{
Hull.White.capletv2(rK, T, Delta = 0.25, ZTO, ZTB, SZTO)
}
\arguments{
\item{rK}{cap rate}

\item{T}{time of caplet payment}

\item{Delta}{time spacing.}

\item{ZTO}{price of a zero coupon bond that matures at time TO using two-factor Vasicek model}

\item{ZTB}{price of a zero coupon bond that matures at time TB using two-factor Vasicek model}

\item{SZTO}{is the appropriate Volatility}
}
\description{
The difference between this function and Hull.White.caplet is that this function uses something similar to 
forward volatilities.
}
