% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hull.White.capv2.R
\name{Hull.White.capv2}
\alias{Hull.White.capv2}
\title{Price of a cap v2}
\usage{
Hull.White.capv2(rK = 0.05, TK = 5, Delta = 0.25, discount_factors, sigmaf)
}
\arguments{
\item{rK}{cap rate}

\item{TK}{maturity of Cap}

\item{Delta}{is normally 0.25}

\item{discount_factors}{are discount factors applicable to 0.25,0.5,.. TK}

\item{sigmaf}{Hull-White forward volatility applicable to each point in Maturity}
}
\description{
The difference between this function and Hull.White.cap is that this function uses something similar to 
forward volatilities.
}
