% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hull.White.put.zcb.R
\name{Hull.White.put.zcb}
\alias{Hull.White.put.zcb}
\title{Price of put option on a zero-coupon bond under Hull-White model.}
\usage{
Hull.White.put.zcb(
  TO = 1,
  TB = 5,
  K = 0.8,
  gamma = 0.019,
  sigma = 0.0196,
  ZCBTB = 0.7978,
  ZCBTO = 0.979
)
}
\arguments{
\item{TO}{option maturity}

\item{TB}{maturity of underlying bond}

\item{K}{strike price of the option}

\item{gamma}{mean reversion rate of Hull-White}

\item{sigma}{instantaneous Hull-White volatility}

\item{ZCBTB}{bond price of a zero coupon that matures at time TB}

\item{ZCBTO}{bond price of a zero coupon that matures at time TO}
}
\description{
Price of put option on a zero-coupon bond under Hull-White model.
}
