% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW.caplet.implied.vol.R
\name{HW.caplet.implied.vol}
\alias{HW.caplet.implied.vol}
\title{This function used in extracting  implied volatility of a caplet.}
\usage{
HW.caplet.implied.vol(
  sigmaf,
  price = 0.0786/100,
  rK = 0.02442,
  T = 0.75,
  Delta = 0.25,
  ZTO = 0.98851,
  ZTB = 0.981899
)
}
\arguments{
\item{sigmaf}{is the implied vol}

\item{price}{of a caplet}

\item{rK}{cap rate applicable to the caplet}

\item{T}{caplet payment tim}

\item{Delta}{the cap rate applicaable is decided at time T-Delta}

\item{ZTO}{discount factor applicable to time T-Delta}

\item{ZTB}{discount faactor applcable to time T}
}
\description{
Its is very similar to Hull.White.caplet.
}
