% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW.forward.vol.R
\name{HW.forward.vol}
\alias{HW.forward.vol}
\title{Calculate forward volatility in Hull-White models.}
\usage{
HW.forward.vol(Maturity, discount_factors, cap_rates, cap_prices)
}
\arguments{
\item{Maturity}{vector containing 0.25,0.5,0.75 etc}

\item{discount_factors}{for times 0.25, 0.5,0.75...}

\item{cap_rates}{cap rates starting at 0.25 going up to the largest cap}

\item{cap_prices}{cap prices for each Maturity}
}
\description{
It uses the function HW.caplet.implied.vol.
}
