% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSdForwards.R
\name{NSdForwards}
\alias{NSdForwards}
\title{First derivative of Nelson-Siegel Forward rates with respect to time.}
\usage{
NSdForwards(coef, maturity)
}
\arguments{
\item{coef}{the vector of parameters \eqn{\beta_0= coef[1], \beta_1=coef[2],\beta_2=coef[3],
\lambda=\frac{1}{\tau}=coef[4]}}

\item{maturity}{is the time to evaluate}
}
\description{
It returns \eqn{\frac{\partial}{\partial t} f(0,t)= -\frac{\beta_1}{\tau}\exp\left(-\frac{t}{\tau}\right)
+\frac{\beta_2}{\tau}\exp\left(-\frac{t}{\tau}\right)-
 \frac{\beta_2t}{\tau^2}\exp\left(-\frac{t}{\tau}\right)}
The package Yield curve express \eqn{\lambda = \frac{1}{\tau}}
}
