% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSForwards.R
\name{NSForwards}
\alias{NSForwards}
\title{Forward rates with Nelson-Siegel Model}
\usage{
NSForwards(coef, maturity)
}
\arguments{
\item{coef}{the vector of parameters \eqn{\beta_0= coef[1], \beta_1=coef[2],\beta_2=coef[3],
\lambda=\frac{1}{\tau}=coef[4]}}

\item{maturity}{is the time to evaluate}
}
\description{
It returns \eqn{f(0,t) = \beta_0 + \beta_1 \exp\left(-\frac{t}{\tau}\right )+
\frac{\beta_2 t}{\tau} \exp\left(-\frac{t}{\tau}\right)}
The package Yield curve express \eqn{\lambda = \frac{1}{\tau}}
}
