% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSRates.R
\name{NSRates}
\alias{NSRates}
\title{Spot rates (yield rate) with Nelson_Siegel method.}
\usage{
NSRates(coef, maturity)
}
\arguments{
\item{coef}{the vector of parameters \eqn{\beta_0= coef[1], \beta_1=coef[2],\beta_2=coef[3],
\lambda=\frac{1}{\tau}=coef[4]}}

\item{maturity}{is the time to evaluate}
}
\description{
It returns \eqn{r(0,t)=\beta_0 + \beta_1 \frac{ 1-\exp\left(-\frac{t}{\tau}\right) }{\frac{t}{\tau}} +\beta_2 \frac{1-\exp\left(-\frac{t}{\tau}\right)-\frac{t}{\tau}\exp\left(-\frac{t}{\tau}\right)}{\frac{t}{\tau}}}
The Yield curve express \eqn{\lambda = \frac{1}{\tau}}
}
