% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfit.R
\name{polyfit}
\alias{polyfit}
\title{Function to evaluate a polynomial or its derivative at a given point.}
\usage{
polyfit(x, coeffs, deriv = 0)
}
\arguments{
\item{x}{the point which polynomial has to be evaluated.}

\item{coeffs}{coefficient vector in the order \eqn{a_0,a_1,\ldots,a_n}}

\item{deriv}{indicates which derivative it needs, calculate up to second derivative}
}
\description{
It returns \eqn{ p(x) = a_0 + a_1 x +a_2x + \ldots+a_nx^n}
}
