% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.zcbp.R
\name{sim.zcbp}
\alias{sim.zcbp}
\title{Zero coupon bond prices using simulation}
\usage{
sim.zcbp(X)
}
\arguments{
\item{X}{is a multivariate time series object. Each column is an interest rate path}
}
\description{
This function takes multiple interest rate paths given as columns of the time series matrix X
Integral is approximated by Riemann-sums
}
