% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theta.R
\name{Theta}
\alias{Theta}
\title{Theta of Hull-White model with Nelson-Siegel curve}
\usage{
Theta(coef, sigma = 0.0221, gamma = 0.19, maturity)
}
\arguments{
\item{coef}{the vector of parameters \eqn{\beta_0= coef[1], \beta_1=coef[2],\beta_2=coef[3],
\lambda=\frac{1}{\tau}=coef[4]}}

\item{sigma}{is \eqn{\sigma}}

\item{gamma}{is \eqn{\gamma}}

\item{maturity}{is the time to evaluate}
}
\description{
It calculates \eqn{\theta_T = \frac{\partial f(0,T)}{\partial T} +
\gamma^* f(0,T)+\frac{\sigma^2}{2\gamma^*}(1-\exp(-2\gamma^*T)}
}
