% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two.factor.Hull.White.cap.price.vector.R
\name{Two.factor.Hull.White.cap.price.vector}
\alias{Two.factor.Hull.White.cap.price.vector}
\title{Calculate cap price vector}
\usage{
Two.factor.Hull.White.cap.price.vector(
  cap_rates,
  Maturity,
  Delta = 0.25,
  discount_factors,
  gamma1,
  gamma2,
  sigma1,
  sigma2,
  rho
)
}
\arguments{
\item{cap_rates}{cap rates starting at 0.25 going upto the largest cap}

\item{Maturity}{vector containing 0.25,0.5,0.75 etc}

\item{Delta}{reciprocal of number of payments per year; it is assumed 4 times payments, Delta =0.25}

\item{discount_factors}{for times 0.25, 0.5,0.75...}

\item{gamma1}{first factor speed of mean reversion}

\item{gamma2}{second factor speed of mean reversion}

\item{sigma1}{first factor  instantaneous  volatility}

\item{sigma2}{second factor instantaneous  volatility}

\item{rho}{instantaneous correlation coefficient between two factors}
}
\description{
This function uses the function Two.factor.Hull.White.cap.
}
\details{
Parameters are identical to that of cap price in the same model.
}
