% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two.factor.Hull.White.cap.R
\name{Two.factor.Hull.White.cap}
\alias{Two.factor.Hull.White.cap}
\title{Price of a cap.}
\usage{
Two.factor.Hull.White.cap(
  rK,
  TK,
  Delta = 0.25,
  discount_factors,
  gamma1,
  gamma2,
  sigma1,
  sigma2,
  rho
)
}
\arguments{
\item{rK}{cap rate}

\item{TK}{is the time of last payment from the cap}

\item{Delta}{reciprocal of number of payments per year; it is assumed 4 times payments, Delta =0.25}

\item{discount_factors}{for times 0.25, 0.5,0.75...}

\item{gamma1}{first factor speed of mean reversion}

\item{gamma2}{second factor speed of mean reversion}

\item{sigma1}{first factor  instantaneous  volatility}

\item{sigma2}{second factor instantaneous  volatility}

\item{rho}{instantaneous correlation coefficient between two factors}
}
\description{
This is obtained by summing up individual caplet prices.
Calculate values of all the caplet that makes payment from 2*Delta to T2 and add them up
}
