% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two.factor.Hull.White.caplet.R
\name{Two.factor.Hull.White.caplet}
\alias{Two.factor.Hull.White.caplet}
\title{Calculates caplet price with the  two-factor Hull-White model.}
\usage{
Two.factor.Hull.White.caplet(
  rK,
  T,
  Delta = 0.25,
  ZTO,
  ZTB,
  gamma1,
  gamma2,
  sigma1,
  sigma2,
  rho
)
}
\arguments{
\item{rK}{cap rate}

\item{T}{time of caplet payment.}

\item{Delta}{time in years between two caplet payments.}

\item{ZTO}{price of a zero coupon bond that matures at time TO using two-factor Vasicek model}

\item{ZTB}{price of a zero coupon bond that matures at time TB using two-factor Vasicek model}

\item{gamma1, gamma2, sigma1, sigma2, rho}{and correlation are \eqn{\gamma_1^*,\gamma_2^*,
\sigma_1,\sigma_2} and \eqn{\rho} respectively.}
}
\description{
Caplet price is calculated as the price of a put option on a zero-coupon bond.
}
