% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two.factor.Vasicek.A.R
\name{Two.factor.Vasicek.A}
\alias{Two.factor.Vasicek.A}
\title{A in the tow-factor Vasicek bond pricing formula.}
\usage{
Two.factor.Vasicek.A(phi1, phi2, gamma1, gamma2, sigma1, sigma2, rho, T)
}
\arguments{
\item{phi1}{long term average for factor 1.}

\item{phi2}{long term average for factor 2.}

\item{gamma1}{first factor speed of mean reversion}

\item{gamma2}{second factor speed of mean reversion}

\item{sigma1}{first factor  instantaneous  volatility}

\item{sigma2}{second factor instantaneous  volatility}

\item{rho}{instantaneous correlation coefficient between two factors}

\item{T}{maturity date.}
}
\description{
This function calculates A as described in equations in the Section 3.6 of the article.
}
\details{
It handles the situation where \eqn{\gamma_1^*=0} and or \eqn{\gamma_2^*=0}
}
