% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two.factor.Vasicek.call.zcb.R
\name{Two.factor.Vasicek.call.zcb}
\alias{Two.factor.Vasicek.call.zcb}
\title{Value of call on a zero-coupon bond.}
\usage{
Two.factor.Vasicek.call.zcb(
  TO,
  TB,
  K,
  ZTO,
  ZTB,
  phi1,
  phi2,
  gamma1,
  gamma2,
  sigma1,
  sigma2,
  rho
)
}
\arguments{
\item{TO}{time to option maturity}

\item{TB}{time to maturity of the underlying zero-coupon bond; when option expires underlying bond will have TB-TO remaining.}

\item{K}{is the strike price.}

\item{ZTO}{price of a zero coupon bond that maturues at time TO using two-factor Vasicek model}

\item{ZTB}{price of a zero coupon bond that matures at time TB using two-factor Vasicek model}

\item{phi1, phi2, gamma1, gamma2, sigma1, sigma2, rho}{are \eqn{\phi_1^*,\phi_2^*,\gamma_1^*,\gamma_2^*,
\sigma_1,\sigma_2} and \eqn{\rho} respectively.}
}
\description{
The call option is  on a zero coupon bond when underlying model is two-factor Vasicek model.
}
