% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two.factor.Vasicek.Rev.Vols.R
\name{Two.factor.Vasicek.Rev.Vols}
\alias{Two.factor.Vasicek.Rev.Vols}
\title{Calculating long and short term volatility.}
\usage{
Two.factor.Vasicek.Rev.Vols(
  gamma1 = 1.7227,
  gamma2 = 0.0434,
  sigma1 = 0.0222,
  sigma2 = 0.0147,
  rho = -0.3969,
  Tau = 5
)
}
\arguments{
\item{gamma1, gamma2, sigma1, sigma2, rho}{and correlation are \eqn{\gamma_1^*,\gamma_2^*,\sigma_1,\sigma_2} and \eqn{\rho} respectively.}

\item{Tau}{time of long rate}
}
\description{
Function to calculate short term volatility, long term volatility and correlation of
the two-factor Vasicek model
It implements some formulas in the Section 3.6 of the article.
}
