% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two.factor.Vasicek.SZ.gradient.R
\name{Two.factor.Vasicek.SZ.gradient}
\alias{Two.factor.Vasicek.SZ.gradient}
\title{Calculate the Jacobian of residual  function.}
\usage{
Two.factor.Vasicek.SZ.gradient(paras, forward_vol.dat)
}
\arguments{
\item{paras}{contains the parameters that need to be estimated
gamma1=paras[1],
gamma2=paras[2],
sigma1 = paras[3],
sigma2 = paras[4],
rho = paras[5].}

\item{forward_vol.dat}{contains maturity and Volatility divided by thee time it calculated}
}
\description{
The residual function is  for forward volatility of two-factor in Hull-White models.
}
