% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two.factor.Vasicek.SZ.R
\name{Two.factor.Vasicek.SZ}
\alias{Two.factor.Vasicek.SZ}
\title{Volatility calculation.}
\usage{
Two.factor.Vasicek.SZ(TO, TB, gamma1, gamma2, sigma1, sigma2, rho)
}
\arguments{
\item{TO}{time to option maturity}

\item{TB}{time to maturity of the underlying zero-coupon bond; when option expires underlying bond will have TB-TO remaining}

\item{gamma1, gamma2, sigma1, sigma2, rho}{are \eqn{\gamma_1^*,\gamma_2^*,
\sigma_1,\sigma_2} and \eqn{\rho} respectively.}
}
\description{
This volatility is used in the call option on bonds when underlying model is the
two-factor Vasicek.
}
