% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two.factor.Vasicek.Trans.sim.R
\name{Two.factor.Vasicek.Trans.sim}
\alias{Two.factor.Vasicek.Trans.sim}
\title{Two-factor Vasicek path simulation using Transition density  approach.}
\usage{
Two.factor.Vasicek.Trans.sim(
  t0 = 0,
  T = 10,
  Delta = 1/252,
  r0 = 0.0437,
  r05 = 0.0454,
  Tau = 5,
  gamma1 = 1.7227,
  gamma2 = 0.04343,
  phi1 = 0.2993,
  phi2 = 0,
  sigma1 = 0.0222,
  sigma2 = 0.0147,
  rho = -0.3969,
  M = 10
)
}
\arguments{
\item{t0}{starting point of interest rate path.}

\item{T}{ending time point.

both t0 and T are given in years.}

\item{Delta}{is the time step; default value being 1 day (1/252 in years).}

\item{r0}{is the starting value of the path (starting value of short rate).}

\item{r05}{starting value of the long term rate.}

\item{Tau}{time in years of long term rate.}

\item{gamma1, gamma2, sigma1, sigma2, rho, phi1, phi2}{are \eqn{\gamma_1^*,\gamma_2^*,
\sigma_1,\sigma_2,\rho, \phi_1} and \eqn{phi_2} respectively.}

\item{M}{is the required number of interest rate paths

Default values of r0, gamma, rbar and alpha are the ones given in Veronesi(2010) Table 15.3 real world parameters.}
}
\description{
This function simulate M  interest rate paths from the Vasicek model using transition density  approach.
}
