% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two.factor.Vasicek.Vols.R
\name{Two.factor.Vasicek.Vols}
\alias{Two.factor.Vasicek.Vols}
\title{This function calculates sigma1, sigma 2 and rho when volatility of short, volatility of long rate and
correlation between long and short in equations are given.}
\usage{
Two.factor.Vasicek.Vols(
  gamma1 = 0.8269,
  gamma2 = -0.0288,
  Tau = 5,
  short.term.vol = 0.0221,
  long.term.vol = 0.0125,
  correlation = 0.4713
)
}
\arguments{
\item{gamma1, gamma2}{are \eqn{\gamma_1^*} and \eqn{\gamma_2^*},respectively.}

\item{Tau}{time in years of long term rate.}

\item{short.term.vol, long.term.vol, correlation}{are volatility of short rate, volatility of long rate and correlation between long rate and short  rate.}
}
\description{
This function calculates sigma1, sigma 2 and rho when volatility of short, volatility of long rate and
correlation between long and short in equations are given.
}
