% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two.factor.Vasicek.yield.R
\name{Two.factor.Vasicek.yield}
\alias{Two.factor.Vasicek.yield}
\title{This function calculates log of  bond prices when the inputs are observed short term volatility, long term volatility}
\usage{
Two.factor.Vasicek.yield(
  t,
  T,
  Tau,
  rt,
  rlTau,
  phi1,
  phi2,
  gamma1,
  gamma2,
  short.term.vol,
  long.term.vol,
  correlation
)
}
\arguments{
\item{t}{current time in years.}

\item{T}{maturity value; the yield is measured at time  for a zero-coupon bond that matures at T.}

\item{Tau}{time in years of long term rate.}

\item{rt}{current short term rate.}

\item{rlTau}{current long term rate.}

\item{gamma1, gamma2, phi1, phi2, }{are \eqn{\gamma_1^*,\gamma_2^*,\phi_1} and \eqn{\phi_2},respectively.}

\item{short.term.vol, long.term.vol, correlation}{are volatility of short rate, volatility of long rate and correlation between long rate and short  rate.
and correlation between short and long term rates.}
}
\description{
This function calculates log of  bond prices when the inputs are observed short term volatility, long term volatility
}
