% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two.factor.Vasicek.zcbp.R
\name{Two.factor.Vasicek.zcbp}
\alias{Two.factor.Vasicek.zcbp}
\title{Zero-coupon bond price under the two-factor Vasicek model.}
\usage{
Two.factor.Vasicek.zcbp(
  t,
  T,
  Tau,
  rt,
  rlTau,
  phi1,
  phi2,
  gamma1,
  gamma2,
  sigma1,
  sigma2,
  rho
)
}
\arguments{
\item{t}{current time in years.}

\item{T}{maturity value}

\item{Tau}{time in years of long term rate.}

\item{rt}{current value of the short rate.}

\item{rlTau}{current  value of the long term rate.}

\item{gamma1, gamma2, sigma1, sigma2, rho, phi1, phi2}{are \eqn{\gamma_1^*,\gamma_2^*,
\sigma_1,\sigma_2,\rho, \phi_1} and \eqn{phi_2} respectively.}
}
\description{
The following function calculates zero-coupon bond prices when short rate rt at time t and 
long rate rlTau at Tau and other parameters
given, it first calculate the short rate factor  phi1t and long rate factor phi2t at time t.
}
