% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two.factor.Vasicek.zcbpv2.R
\name{Two.factor.Vasicek.zcbpv2}
\alias{Two.factor.Vasicek.zcbpv2}
\title{Zero-coupon bond price under the two-factor Vasicek model.}
\usage{
Two.factor.Vasicek.zcbpv2(
  t,
  T,
  Tau,
  rt,
  rlTau,
  phi1,
  phi2,
  gamma1,
  gamma2,
  short.term.vol,
  long.term.vol,
  correlation
)
}
\arguments{
\item{t}{time of bond price calculation}

\item{T}{bond maturity time}

\item{Tau}{time of long rate.}

\item{rt}{short rate}

\item{rlTau}{long rate}

\item{phi1, phi2, gamma1, gamma2}{are \eqn{\phi_1^*,\phi_2^*,\gamma_1^*,\gamma_2^*} and \eqn{\rho} respectively.\\}

\item{short.term.vol, long.term.vol, correlation}{are observed short term volatility, long term volatility
and correlation between short and long term rate}
}
\description{
This function calculates log of  bond prices when the inputs are observed short term volatility, long term volatility
and correlation between short and long term rates.
}
