% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vasicek.B.R
\name{Vasicek.B}
\alias{Vasicek.B}
\title{B(0;T) in the Vasicek bond pricing formula.}
\usage{
Vasicek.B(gamma, T)
}
\arguments{
\item{gamma}{is the mean reversion speed}

\item{T}{is the time to maturity}
}
\description{
It returns \deqn{B(0,T)=\frac{1-\exp(-\gamma T)}{\gamma},~\text{if}~~ \gamma \ne =0}
\deqn{B(0,T)= T~\text{if}~~ \gamma  =0}
}
