% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vasicek.call.delta.R
\name{Vasicek.call.delta}
\alias{Vasicek.call.delta}
\title{Function to calculate DELTA of a call option}
\usage{
Vasicek.call.delta(
  r0 = 0.03,
  TO = 0.5,
  TB = 1,
  K = 0.98,
  gamma = 0.4653,
  rbar = 0.0634,
  sigma = 0.0221
)
}
\arguments{
\item{r0}{short  rate at time 0}

\item{TO}{option maturity time}

\item{TB}{zero coupon bond maturity time}

\item{K}{option strike price as a percentage of face value}

\item{gamma}{is speed of mean reversion}

\item{rbar}{is the long term mean.}

\item{sigma}{is the instantaneous volatility.}
}
\description{
Function to calculate DELTA of a call option
}
