% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vasicek.call.zcb.R
\name{Vasicek.call.zcb}
\alias{Vasicek.call.zcb}
\title{Price of a call option on zero coupon bond with Vasicek model}
\usage{
Vasicek.call.zcb(
  r0 = 0.03,
  TO = 0.5,
  TB = 1,
  K = 0.98,
  gamma = 0.4653,
  rbar = 0.0634,
  sigma = 0.0221
)
}
\arguments{
\item{r0}{short  rate at time 0}

\item{TO}{option maturity time}

\item{TB}{zero coupon bond maturity time}

\item{K}{option strike price as a percentage of face value}

\item{gamma}{is speed of mean reversion}

\item{rbar}{is the long term mean.}

\item{sigma}{is the instantaneous volatility.
default \eqn{\gamma,\bar{r}} and \eqn{\sigma} are the values in example 15.1}
}
\description{
It implements European call price formula 
\deqn{V(r_0,0) = Z(r_0,0;T_B){N}(d_1) -KZ(r_0,0;T_O){N}(d_2)}
\deqn{d_1 = \frac{1}{S_Z(T_O;T_B)} \log \left(\frac{Z(r_0,0;T_B)}{KZ(r_0,0;T_O)}\right)+\frac{S_Z(T_O;T_B)}{2}}
\deqn{d_2 = d_1 - S_Z(T_O;T_B)}
\deqn{S_Z(T_O;T_B) = B(T_O;T_B) \times \sqrt{\frac{\sigma^2}{2\gamma^*}(1 - e^{-2\gamma^* T_O})}}
}
