% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vasicek.cbp.R
\name{Vasicek.cbp}
\alias{Vasicek.cbp}
\title{Vasicek formula for coupon bonds}
\usage{
Vasicek.cbp(
  rTO = 0.0342,
  gamma = 0.4653,
  rbar = 0.0634,
  sigma = 0.0221,
  TO = 1,
  cash_flow_timing = c(3:10)/2,
  cash_flow = c(rep(2.5, 7), 102.5)
)
}
\arguments{
\item{rTO}{is the short rate at time t}

\item{gamma}{is the mean reversion speed}

\item{rbar}{is the long term mean}

\item{sigma}{is the instantaneous volatility.}

\item{TO}{the point of valuation}

\item{cash_flow_timing}{vector containing cash flow timing in years}

\item{cash_flow}{vector containing cash flows at cash_flow_timing}
}
\description{
Vasicek formula for coupon bonds
}
