% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vasicek.DB.R
\name{Vasicek.DB}
\alias{Vasicek.DB}
\title{Derivative of B (Vasicek bond pricing formula B)}
\usage{
Vasicek.DB(gamma, T)
}
\arguments{
\item{gamma}{is the mean reversion speed}

\item{T}{is the time to maturity}
}
\description{
It returns derivative of \eqn{B=(0,T)=\frac{1-\exp(-\gamma T)}{\gamma}} with respect to \eqn{\gamma}
the derivative at  \eqn{gamma=0} calculated as the limit.
}
