% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vasicek.delta.hedge.R
\name{Vasicek.delta.hedge}
\alias{Vasicek.delta.hedge}
\title{delta hedging of zero coupon bond}
\usage{
Vasicek.delta.hedge(
  r0 = 0.0168,
  T1 = 1.1041,
  T2 = 3.3479,
  delta = 1/252,
  gamma = 0.3261,
  rbar = 0.0509,
  gammast = 0.4653,
  rbarst = 0.0634,
  sigma = 0.0221
)
}
\arguments{
\item{r0}{initial rate}

\item{T1}{maturity of short bond}

\item{T2}{maturity of long bond}

\item{delta}{portfolio rebalancing time intervals}

\item{gamma}{is speed of mean reversion in real world}

\item{rbar}{is the long term mean in real world.}

\item{gammast}{is speed of mean reversion in risk neutral world}

\item{rbarst}{is the long term mean in risk neutral world.}

\item{sigma}{is the instantaneous volatility.}
}
\description{
delta hedging of zero coupon bond
}
