% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vasicek.J.R
\name{Vasicek.J}
\alias{Vasicek.J}
\title{Function to compute minimizing quantity given in section 3.2
\eqn{J(\gamma^*,\bar{r}^*) = \sum_{i=1}^n (Z^{Vasicek}(r_0,0;T_i) - Z^{Data}(0,T_i))^2}}
\usage{
Vasicek.J(param, r0 = 0.0168, sigma = NULL, bond.prices, bond.maturities)
}
\arguments{
\item{param}{is a list of parameters c(\eqn{\gamma,\bar{r},\sigma}).}

\item{r0}{initial short rate default value is 0.0168}

\item{sigma}{if it is set to null minimization is  with respct to all three parameters \eqn{\gamma,\alpha} and \eqn{\sigma} otherwise
its minimiztion is with respect to \eqn{\gamma} and \eqn{\bar{r}} only.}

\item{bond.prices}{a list of observed zero coupon bond prices (or discount factors) for bond.maturities}

\item{bond.maturities}{a list bond maturities}
}
\description{
This function will be used in ``nlm" and ``optim" as the objective function to be minimized.
It uses the function Vasicek.zcbp
}
