% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vasicek.sim.call.zcb.R
\name{Vasicek.sim.call.zcb}
\alias{Vasicek.sim.call.zcb}
\title{Call option price based on brute force simulation with the Vasicek model}
\usage{
Vasicek.sim.call.zcb(
  r0 = 0.03,
  TO = 0.5,
  TB = 1,
  K = 0.98,
  gamma = 0.4653,
  rbar = 0.0634,
  sigma = 0.0221,
  Delta = 1/252,
  M = 30
)
}
\arguments{
\item{r0}{short  rate at time 0}

\item{TO}{option maturity time}

\item{TB}{zero coupon bond maturity time}

\item{K}{option strike price as a percentage of face value}

\item{gamma}{is speed of mean reversion}

\item{rbar}{is the long term mean.}

\item{sigma}{is the instantaneous volatility.

default \eqn{\gamma,\bar{r}} and \eqn{\sigma} are the values in example 15.1
default Delta is one day in years}

\item{Delta}{time step for simulation}

\item{M}{is the number of paths to be simulated}
}
\description{
It uses transition density method of simulating interest rate paths.
}
