% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vasicek.SZ.R
\name{Vasicek.SZ}
\alias{Vasicek.SZ}
\title{This is the volatility used in Vasicek, Ho-Lee and Hull-White option pricing formulas.}
\usage{
Vasicek.SZ(TO = 0.5, TB = 1, gamma = 0.4653, sigma = 0.0221)
}
\arguments{
\item{TO}{option maturity time}

\item{TB}{zero coupon bond maturity time}

\item{gamma}{is speed of mean reversion}

\item{sigma}{is the instantaneous volatility.
default \eqn{\gamma,\bar{r}} and \eqn{\sigma} are the values in example 15.1}
}
\description{
It implements the following formula which can be used in many routines.
\deqn{S_Z(T_O;T_B) = B(T_O;T_B) \times \sqrt{\frac{\sigma^2}{2\gamma^*}(1 - e^{-2\gamma^* T_O})}}
\deqn{S_Z(T_O;T_B) = \sigma (T_B-T_O)\sqrt{T_O}~~\text{if} \gamma^*= 0}
}
