% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vasicek.zcbp.R
\name{Vasicek.zcbp}
\alias{Vasicek.zcbp}
\title{Vasicek zero coupon bond prices}
\usage{
Vasicek.zcbp(
  r0 = 0.03,
  t = 0,
  T = 1,
  gamma = 0.4653,
  rbar = 0.0634,
  sigma = 0.0221
)
}
\arguments{
\item{r0}{is the starting value of the path.}

\item{t}{starting point of interest rate path.}

\item{T}{ending time point.

both t and T are given in years.}

\item{gamma}{is speed of mean reversion}

\item{rbar}{is the long term mean.}

\item{sigma}{is the instantaneous volatility.}
}
\description{
This function calculates zero coupon bond prices when short rates follow the Vasicek model.
It Implements the following formulas.
\deqn{B(t;T) = \frac{1}{\gamma^*}\left( 1 - e^{-\gamma^* (T-t)}\right)}
\deqn{A(t;T) = (B(t;T) - (T-t))\left( \bar{r}^* - \frac{\sigma^2}{2(\gamma^*)^2} \right) - \frac{\sigma^2 B(t;T)^2}{4\gamma^*}}
\deqn{Z(r,t;T) = e^{A(t;T)-B(t;T) r}}
default parameter values are those of Veronesi (2010) Example 15.1
}
