% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vasicek.zcbp1.R
\name{Vasicek.zcbp1}
\alias{Vasicek.zcbp1}
\title{zero coupon bond pricing formula}
\usage{
Vasicek.zcbp1(
  r0 = 0.03,
  t = 0,
  T = 1,
  gamma = 0.4653,
  rbar = 0.0634,
  sigma = 0.0221
)
}
\arguments{
\item{r0}{is the starting value of the path.}

\item{t}{starting point of interest rate path.}

\item{T}{ending time point.

both t and T are given in years.}

\item{gamma}{is speed of mean reversion}

\item{rbar}{is the long term mean.}

\item{sigma}{is the instantaneous volatility.}
}
\description{
This is similar to Vasicek.zcbp but returning  additional intermediate values
}
\details{
default parameter values are those of Veronesi (2010) Example 15.1
}
